move handler registration back to room
